+++ src/resources/lang/zh-cn/menus.php
@@ -16,7 +16,6 @@ return array(
     'news' => '新闻管理',
     'portfolios' => '作品管理',
     'purchases' => '购买历史',
-    'discounts' => '折扣管理',
     'mailinglist' => '群发邮件',
     'coupons' => '优惠券管理'
 );