+++ src/resources/lang/zh-tw/menus.php
@@ -16,7 +16,6 @@ return array(
     'news' => '新聞管理',
     'portfolios' => '作品管理',
     'purchases' => '購買歷史',
-    'discounts' => '折扣管理',
     'mailinglist' => '群發郵件',
     'coupons' => '優惠券管理'
 );